+++ crates/nwc/src/prelude.rs
@@ -11,4 +11,6 @@
 pub use nostr::prelude::*;
 pub use nostr_relay_pool::prelude::*;
 
+pub use crate::builder::*;
+pub use crate::error::*;
 pub use crate::*;